feat: Client => Add Dropdown component, improve design
